Moved methods into imports
To clean up some code and move from addFiles to mainModule methods were
moved into imports
