--- logglyMeteorMethods.js
@@ -1,32 +1,0 @@
-Meteor.methods({
-  logglyLog: function(param, tag) {
-    check(param, Match.OneOf(String, Object));
-    check(tag, Match.OneOf([String], null));
-
-    Logger.log(param, tag);
-  },
-  logglyTrace: function(param, tag) {
-    check(param, Match.OneOf(String, Object));
-    check(tag, Match.OneOf([String], null));
-
-    Logger.trace(param, tag);
-  },
-  logglyInfo: function(param, tag) {
-    check(param, Match.OneOf(String, Object));
-    check(tag, Match.OneOf([String], null));
-
-    Logger.info(param, tag);
-  },
-  logglyWarn: function(param, tag) {
-    check(param, Match.OneOf(String, Object));
-    check(tag, Match.OneOf([String], null));
-
-    Logger.warn(param, tag);
-  },
-  logglyError: function(param, tag) {
-    check(param, Match.OneOf(String, Object));
-    check(tag, Match.OneOf([String], null));
-
-    Logger.error(param, tag);
-  }
-});
