+++ package.js
@@ -12,10 +12,9 @@ Npm.depends({
 Package.onUse(function (api) {
   api.versionsFrom('0.9.4');
   api.use(['check']);
-  api.addFiles('logglyServer.js', 'server');
-  api.addFiles('logglyMeteorMethods.js', 'server');
-  api.addFiles('logglyClient.js', 'client');
   api.export('Loggly', 'server');
+  api.mainModule('logglyClient.js', 'client');
+  api.mainModule('logglyServer.js', 'server');
   api.export('Logger', 'client');
   api.export('Logger', 'server'); // Normally overwritten by users but handles cases of undefined Loggers on the server.
 });
